Added timeout parameter to send_ws_request
The requests module that the send_ws_request relies on has an optional timeout parameter that will make sure that the request doesn't go for longer than the timeout value (in seconds).

I have added an optional timeout param on the send_ws_request function to allow users of this module the ability to limit how long they are willing to wait for responses.

A requests.errors.Timeout exception will be raised it the timeout is not None and reached. It is simple enough for users of the byu_ws_sdk package to handle this exception if they are passing the timeout parameter.