Added `authorize_request` function.
The `authorize_request` function takes the requested url and the value
given in the Authorization header, among several other arguments, and
determines if the request was properly authorized. Returns the personId
of the valid authorization or None if not properly authorized.

This function could be used in decorators or just by calling from web
service code to determine if the requester is authorized to make
requests to your service. The valid personId can be checked against an
authorized personId list or just if it isn't None to allow all
authorized callers from OIT/BYU.

Note: This method works great for me and the way I need to check
requests. I don't know a ton about all the possible ways of
authorization checking, so this might not work for all cases.

Also added two tests for this function. One that should fail and one
that should success, given correct credentials on test execution.