Fixed the input tests for python3.
This was a relatively simple fix. Because the raw_input function doesn't
exist in python3, we had to just do a check to get the right 'input'
function to use.

For me these input tests give ResourceWarnings in python3. These
warnings don't seem to interfere with the tests success or failure, they
just deal with the use of the input and getpass functions in python3.

For me the test_authorize_request test doesn't work in python3. This is
using the exact same credentials as I do when running it in python2, but
they fail in python3. I'm wondering if it had to do with the different
code running for the different versions in the base64encode_string and
make_sha512_mac functions in core.py. I'm looking for confirmation on
this. I'm not sure what these functions are supposed to do.